1. move remote module utils to downloader pkg 2. add wiki link for punycode 3. use filepath.Join()
