+++ pkg/downloader/remote-modules.go
@@ -14,7 +14,7 @@
     limitations under the License.
 */
 
-package utils
+package downloader
 
 import (
 	"strings"
